# gpt-assistant-openapi.yaml
openapi: 3.0.0
info:
  title: Driver Route Optimization API
  description: Comprehensive API for GPT assistant to manage driver route assignments with intelligent OR Tools optimization
  version: 1.0.0
  contact:
    name: Route Optimization System
    description: Intelligent driver assignment with monthly balance optimization
servers:
  - url: https://your-app.replit.app/api
    description: Production server (replace with your actual Replit URL)
  - url: http://localhost:5000/api
    description: Development server
paths:
  /ortools/solve:
    post:
      summary: Solve Optimal Route Assignment
      description: Uses OR Tools to find optimal driver-route assignments considering monthly hour balance
      operationId: solveOptimalAssignment
      responses:
        '200':
          description: Optimization completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResult'
        '500':
          description: Optimization failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ortools/apply:
    post:
      summary: Apply Optimization Results
      description: Saves the optimized assignments to database and updates driver hours
      operationId: applyOptimization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                solution:
                  $ref: '#/components/schemas/OptimizationResult'
                assignedDate:
                  type: string
                  format: date
                  description: Date for the assignments (YYYY-MM-DD)
              required:
                - solution
                - assignedDate
      responses:
        '200':
          description: Solution applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Solution applied successfully"
        '400':
          description: Invalid request
        '500':
          description: Application failed
...
components:
  schemas:
    OptimizationResult:
      type: object
      properties:
        status:
          type: string
          enum: [optimal, infeasible, error]
          description: Status of the optimization
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/RouteAssignment'
        driver_status:
          type: array
          items:
            $ref: '#/components/schemas/DriverStatus'
        unassigned_routes:
          type: array
          items:
            type: string
        statistics:
          $ref: '#/components/schemas/OptimizationStats'
        objective_value:
          type: number
          description: Total hours assigned (optimization objective)
    RouteAssignment:
      type: object
      properties:
        driver_name:
          type: string
          description: Name of the assigned driver
        route_name:
          type: string
          description: Name of the assigned route
        route_hours:
          type: number
          description: Hours required for this route
    DriverStatus:
      type: object
      properties:
        name:
          type: string
          description: Driver name
        assigned_route:
          type: string
          nullable: true
          description: Route assigned to this driver (null if not assigned)
        assigned_hours:
          type: number
          description: Hours assigned to this driver
        remaining_hours:
          type: number
          description: Monthly hours remaining after assignment
    OptimizationStats:
      type: object
      properties:
        total_routes:
          type: integer
          description: Total number of routes
        routes_assigned:
          type: integer
          description: Number of routes successfully assigned
        routes_unassigned:
          type: integer
          description: Number of routes not assigned
        total_hours_assigned:
          type: number
          description: Total hours assigned across all routes
        drivers_working:
          type: integer
          description: Number of drivers with assignments
        drivers_available:
          type: integer
          description: Total number of available drivers
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information

# GPT Assistant Integration Examples
...